refactor: migrate book state management to Context API - Created BookContext and BookProvider to encapsulate book-related state and logic - Defined BookContextType with books, errorMsg, and async CRUD functions - Replaced local state in App.tsx with useBooks hook from BookContext - Moved fetchBooks into BookProvider to ensure initial data is loaded centrally - Exported useBooks hook for simplified access to context in components - Updated main.tsx to wrap App with BookProvider - Simplified App.tsx to focus only on consuming context data and functions
