+++ books/src/types/BookContextType.ts
@@ -1,0 +1,9 @@
+import type { Book } from "./bookTypes";
+
+export type BookContextType = {
+  books: Book[];
+  errorMsg: string | null;
+  createBook: (title: string) => Promise<void>;
+  deleteBookById: (id: number) => Promise<void>;
+  editBookById: (id: number, newtitle: string) => Promise<void>;
+};
